requirements.txt
google-genai
pillow